Refactored context into new module.
