+++ test/test_templating.py
@@ -10,6 +10,7 @@ from pytest import param
 
 from generate_changelog import configuration, templating
 from generate_changelog.configuration import DEFAULT_SECTION_PATTERNS, get_config
+from generate_changelog.context import CommitContext
 
 FIXTURES_DIR = Path(__file__).parent / "fixtures"
 fake = Faker()
@@ -39,7 +40,7 @@ def test_first_matching(string, expected):
 def test_commit_context():
     """CommitContexts should properly parse things."""
     commit = conftest.commit_factory()
-    context = templating.CommitContext(
+    context = CommitContext(
         sha=commit.hexsha,
         commit_datetime=commit.committed_datetime,
         committer=f"{commit.committer.name} <{commit.committer.email}>",
@@ -56,7 +57,7 @@ def test_commit_with_no_email():
     """A trailer without an email should still get parsed."""
     commit = conftest.commit_factory()
     name_only = fake.name()
-    context = templating.CommitContext(
+    context = CommitContext(
         sha=commit.hexsha,
         commit_datetime=commit.committed_datetime,
         committer=f"{commit.committer.name} <{commit.committer.email}>",
